requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
sqlalchemy>=2.0.37
httpx>=0.28.0
pyyaml>=6.0.1
python-multipart>=0.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openmemory>=0.1.0
aiofiles>=24.1.0
typer>=0.15.0
rich>=13.0.0
psutil>=6.0.0
